Revert "Prevent errors from unsupported plugins/rules/modules (#215)"
This reverts commit 55c8649071820774cdf4ecc0905ee63521334040.
