+++ Makefile
@@ -1,24 +1,11 @@
-.PHONY: image test citest integration
+.PHONY: image test citest
 
 IMAGE_NAME ?= codeclimate/codeclimate-eslint
 
 NPM_TEST_TARGET ?= test
-NPM_INTEGRATION_TARGET ?= integration
-
-DEBUG ?= false
-ifeq ($(DEBUG),true)
-	NPM_TEST_TARGET = test.debug
-	NPM_INTEGRATION_TARGET = integration.debug
-endif
 
 image:
 	docker build --rm -t $(IMAGE_NAME) .
-
-integration: image
-	docker run -ti --rm \
-		--volume $(PWD):/code \
-		--workdir /code \
-		$(IMAGE_NAME) npm run $(NPM_INTEGRATION_TARGET)
 
 test: image
 	docker run -ti --rm \
@@ -29,4 +16,4 @@ test: image
 citest:
 	docker run --rm \
 		--workdir /usr/src/app \
-		$(IMAGE_NAME) sh -c "npm run test && npm run integration"
+		$(IMAGE_NAME) npm run test
