--- integration/empty_config/index.js
@@ -1,2 +1,0 @@
-function dummy() {
-}
