--- integration/extends_airbnb/config.json
@@ -1,10 +1,0 @@
-{
-  "enabled": true,
-  "config": {
-    "config": "extends_airbnb/eslintrc.json",
-    "debug": "true"
-  },
-  "include_paths": [
-    "/usr/src/app/integration/extends_airbnb/index.js"
-  ]
-}
