--- integration/extends_airbnb/eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "extends": "airbnb",
-  "parser": "babel-eslint",
-  "rules": {}
-}
