--- integration/with_unsupported_plugins/config.json
@@ -1,10 +1,0 @@
-{
-  "enabled": true,
-  "config": {
-    "config": "with_unsupported_plugins/eslintrc.yml",
-    "debug": "false"
-  },
-  "include_paths": [
-    "/usr/src/app/integration/with_unsupported_plugins/index.js"
-  ]
-}
