--- integration/with_unsupported_plugins/eslintrc.yml
@@ -1,21 +1,0 @@
-env:
-    es6: true
-    node: true
-parserOptions:
-    sourceType: module
-plugins:
-    - node
-    - not_supported
-extends:
-    - not_valid
-    - 'plugin:invalidplugin/recommended'
-    - 'eslint:recommended'
-    - 'plugin:node/recommended'
-rules:
-    invalidplugin/rule: 1
-    node/exports-style: [error, module.exports]
-    indent: [error, 4]
-    linebreak-style: [error, unix]
-    quotes: [error, double]
-    semi: [error, always]
-    no-console: off
