--- integration/with_unsupported_plugins/index.js
@@ -1,2 +1,0 @@
-function dummy() {
-}
