+++ lib/docs.js
@@ -7,26 +7,34 @@
 var fs = require('fs')
   , path = require('path');
 
-function Docs() {
+//------------------------------------------------------------------------------
+// Privates
+//------------------------------------------------------------------------------
 
-  var docs = {};
+var docs = Object.create(null);
 
-  function get(ruleId) {
-    return docs[ruleId];
-  }
-
+function load() {
   var docsDir = path.join(__dirname, '/docs/rules');
 
-  fs.existsSync(docsDir) && fs.readdirSync(docsDir).forEach(function(file) {
+  fs.readdirSync(docsDir).forEach(function(file) {
     var content = fs.readFileSync(docsDir + '/' + file, 'utf8');
 
     // Remove the .md extension from the filename
     docs[file.slice(0, -3)] = content;
   });
-
-  return {
-    get: get
-  };
 }
 
-module.exports = Docs;
+//------------------------------------------------------------------------------
+// Public Interface
+//------------------------------------------------------------------------------
+
+exports.get = function(ruleId) {
+  return docs[ruleId];
+};
+
+//------------------------------------------------------------------------------
+// Initialization
+//------------------------------------------------------------------------------
+
+// loads existing docs
+load();
