--- lib/empty-plugin.js
@@ -1,6 +1,0 @@
-module.exports = {
-  rules: {},
-  configs: {
-    recommended: {}
-  }
-};
