+++ lib/eslint-patch.js
@@ -1,43 +1,33 @@
-"use strict";
+'use strict';
+var meld = require('meld');
+var docs = require('./docs');
+var Config = require("eslint/lib/config");
+var ConfigUpgrader = require('./config_upgrader');
 
-const Plugins = require("eslint/lib/config/plugins");
-const ModuleResolver = require("eslint/lib/util/module-resolver");
+var supportedPlugins = ['react', 'babel'];
 
-const ConfigFile = require("eslint/lib/config/config-file");
+module.exports = function patcher(eslint) {
 
-const Config = require("eslint/lib/config");
-const ConfigUpgrader = require("./config_upgrader");
+  meld.around(eslint.CLIEngine, 'loadPlugins', function(joinPoint) {
+    var pluginNames = joinPoint.args[0];
+    var filteredPluginNames = pluginNames.filter(function(pluginName) {
+      return supportedPlugins.indexOf(pluginName) >= 0;
+    });
+    return joinPoint.proceed(filteredPluginNames);
+  });
 
-module.exports = function patch() {
-  const skippedModules = [];
-  function warnModuleNotSupported(name) {
-    if(!skippedModules.includes(name)) {
-      skippedModules.push(name);
-      console.error(`Module not supported: ${name}`);
-    }
-  }
+  meld.around(eslint.CLIEngine, 'addPlugin', function() {
+    return;
+  });
 
-  const resolve = ModuleResolver.prototype.resolve;
-  ModuleResolver.prototype.resolve = function(name, path) {
-    try {
-      return resolve.apply(this, [name, path]);
-    } catch(e) {
-      warnModuleNotSupported(name);
-      return `${__dirname}/empty-plugin.js`;
-    }
-  };
-
-  Plugins.loadAll = function(pluginNames) {
-    for(const name of pluginNames) {
-      try {
-
-        Plugins.load(name);
-
-      } catch(e) {
-        warnModuleNotSupported(`eslint-plugin-${name}`);
-      }
-    }
-  };
+  // meld.around(eslint.CLIEngine.Config, 'loadPackage', function(joinPoint) {
+  //   var filePath = joinPoint.args[0];
+  //   if (filePath.match(/^eslint-config-airbnb.*/)) {
+  //     return joinPoint.proceed();
+  //   } else {
+  //     return {};
+  //   }
+  // });
 
   const originalGetConfig = Config.prototype.getConfig;
   Config.prototype.getConfig = function(filePath) {
@@ -47,5 +37,7 @@ module.exports = function patch() {
     return configUpgrader.upgrade(originalConfig);
   };
 
-  return require('eslint');
+  eslint.docs = docs;
+
+  return eslint;
 };
