+++ package.json
@@ -59,7 +59,8 @@
     "eslint-plugin-vue": "^2.0.1",
     "eslint-plugin-xogroup": "^1.2.0",
     "glob": "^7.0.6",
-    "prettier": "^0.22.0"
+    "prettier": "^0.22.0",
+    "meld": "^1.3.2"
   },
   "devDependencies": {
     "chai": "^3.5.0",
@@ -69,10 +70,8 @@
     "temp": "^0.8.3"
   },
   "scripts": {
-    "integration": "mocha -gc integration",
-    "integration.debug": "mocha -gc debug integration",
-    "test": "mocha -gc test",
-    "test.debug": "mocha -gc debug test"
+    "test": "mocha test",
+    "test.debug": "mocha debug test"
   },
   "engine": "node >= 0.12.4"
 }
